Update sunburst visualization with Nashville data and labeling
Nashville Data and Whitelabeling added to the main sunburst visualization. The visualization is complete and works - allowing you to switch between 2014 and 2015, and actual and expected budget data.

We may update to exclude the school system budget, layer differently, or allow users to zoom in and out. Explanations and timelines may also be included.